requirements.txt
cobs
device-communication
libscrc
protobuf==6.*
pyserial
